stockanomalydetectionapp/requirements.txt
quixstreams==3.13.1
python-dotenv
tqdm
pandas
zstandard
scikit-learn